pro :: navigating 세팅
